WIN386 parameter type fixes.
[git-p4: depot-paths = "//depot/openwatcom/": change = 24621]

+++ bld/clib/win386/c/enumobj.c
@@ -32,7 +32,7 @@
 
 #include "cover.h"
 
-short PASCAL _Cover_EnumObjects( HDC dc, int obj, FARPROC p, LPSTR data )
+short PASCAL _Cover_EnumObjects( HDC dc, short obj, FARPROC p, LPSTR data )
 {
     return( EnumObjects( dc, obj, SetProc( p, GETPROC_ENUMOBJECTS ),
                 (LPARAM)data ) );
